+++ NEWS.md
@@ -14,11 +14,13 @@
 
 * Both `pad` and `thicken` will no longer break when there are missing values in the datetime variable. Rows containing missing values will be retained in the returned data frame. In the case of `thicken` they will remain on the same position as the input data frame. The added column will have a missing value as well. For `pad` all the rows with missing values will be moved to the end of the dataframe, since there is no natural position for them in the order of padded rows.
 
-## Bug Fixes
+## Bug Fixes / Enhancements
 
 * When time variable has NULL as timezone, also `posix_to_date` used to break (related to #14). This made `thicken` break when the desired interval is "day" or higher. This is now fixed by don't regarding the timezone.
 
 * `get_interval` now throws an informative error when the datetime variable has missing values (#33).
+
+* `pad` now throws an informative error when the datetime variable is used in the grouping (#38)
 
 ## Further Changes